+++ tests/test_unit/test_throttling/test_throttling_reports.py
@@ -17,8 +17,9 @@ from dmr.plugins.pydantic import PydanticSerializer
 from dmr.renderers import Renderer
 from dmr.test import DMRAsyncRequestFactory, DMRRequestFactory
 from dmr.throttling import AsyncThrottle, Rate, SyncThrottle, ThrottlingReport
+from dmr.throttling.algorithms import LeakyBucket
 from dmr.throttling.cache_keys import RemoteAddr
-from dmr.throttling.headers import RateLimitIETFDraft, XRateLimit
+from dmr.throttling.headers import RateLimitIETFDraft, RetryAfter, XRateLimit
 
 _draft_headers: Final = RateLimitIETFDraft()
 _ratelimit_headers: Final = XRateLimit()
@@ -48,6 +49,7 @@ class _ReportsController(Controller[PydanticSerializer]):
 
 def test_throttle_multiple_headers(
     dmr_rf: DMRRequestFactory,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Ensures that throttle information can be served on success."""
     request = dmr_rf.get('/whatever/')
@@ -96,6 +98,7 @@ class _AsyncReportsController(Controller[PydanticSerializer]):
 @pytest.mark.asyncio
 async def test_throttle_multiple_headers_async(
     dmr_async_rf: DMRAsyncRequestFactory,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Ensures that throttle information can be served on success."""
     request = dmr_async_rf.get('/whatever/')
@@ -148,6 +151,7 @@ class _MultipleThrottlesController(Controller[PydanticSerializer]):
 
 def test_throttle_multiple_throttles(
     dmr_rf: DMRRequestFactory,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Ensures that throttle information can be served on success."""
     request = dmr_rf.get('/whatever/')
@@ -196,6 +200,7 @@ class _AsyncMultipleThrottlesController(Controller[PydanticSerializer]):
 @pytest.mark.asyncio
 async def test_throttle_multiple_throttles_async(
     dmr_async_rf: DMRAsyncRequestFactory,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Ensures that throttle information can be served on success."""
     request = dmr_async_rf.get('/whatever/')
@@ -359,6 +364,7 @@ class _NoThrottleAsyncController(Controller[PydanticSerializer]):
 @pytest.mark.asyncio
 async def test_no_throttle_report_async(
     dmr_async_rf: DMRAsyncRequestFactory,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Ensures that no throttle produces empty reports."""
     request = dmr_async_rf.get('/whatever/')
@@ -370,4 +376,64 @@ async def test_no_throttle_report_async(
     assert isinstance(response, HttpResponse)
     assert response.status_code == HTTPStatus.OK, response.content
     assert response.headers == {'Content-Type': 'application/json'}
+    assert json.loads(response.content) == 'inside'
+
+
+_retry_after: Final = RetryAfter()
+
+
+class _AsyncLeakyBucketController(Controller[PydanticSerializer]):
+    @validate(
+        ResponseSpec(
+            str,
+            status_code=HTTPStatus.OK,
+            headers={
+                **_draft_headers.provide_headers_specs(),
+                **_retry_after.provide_headers_specs(),
+            },
+        ),
+        throttling=[
+            AsyncThrottle(
+                1,
+                Rate.second,
+                response_headers=[_draft_headers, _retry_after],
+                cache_key=RemoteAddr(name='one'),
+                algorithm=LeakyBucket(),
+            ),
+            AsyncThrottle(
+                5,
+                Rate.minute,
+                response_headers=[_draft_headers],
+                cache_key=RemoteAddr(name='two'),
+                algorithm=LeakyBucket(),
+            ),
+        ],
+    )
+    async def get(self) -> HttpResponse:
+        return self.to_response(
+            'inside',
+            headers=await ThrottlingReport(self).areport(),
+        )
+
+
+@pytest.mark.asyncio
+async def test_leaky_bucket_async(
+    dmr_async_rf: DMRAsyncRequestFactory,
+    freezer: FrozenDateTimeFactory,
+) -> None:
+    """Ensures that async throttle reports are correct for leaky bucket algo."""
+    request = dmr_async_rf.get('/whatever/')
+
+    response = await dmr_async_rf.wrap(
+        _AsyncLeakyBucketController.as_view()(request),
+    )
+
+    assert isinstance(response, HttpResponse)
+    assert response.status_code == HTTPStatus.OK, response.content
+    assert response.headers == {
+        'RateLimit-Policy': '1;w=1;name="one", 5;w=60;name="two"',
+        'RateLimit': '"one";r=0;t=1, "two";r=4;t=12',
+        'Retry-After': '1',
+        'Content-Type': 'application/json',
+    }
     assert json.loads(response.content) == 'inside'